refactor(core): Move `DbConnection` to `@n8n/db`
