+++ packages/@n8n/db/package.json
@@ -25,6 +25,7 @@
     "@n8n/backend-common": "workspace:^",
     "@n8n/config": "workspace:^",
     "@n8n/constants": "workspace:^",
+    "@n8n/decorators": "workspace:^",
     "@n8n/di": "workspace:^",
     "@n8n/permissions": "workspace:^",
     "@n8n/typeorm": "catalog:",
