+++ packages/@n8n/db/src/index.ts
@@ -28,3 +28,5 @@ export { postgresMigrations } from './migrations/postgresdb';
 
 export { wrapMigration } from './migrations/migration-helpers';
 export * from './migrations/migration-types';
+export { DbConnection } from './connection/db-connection';
+export { DbConnectionOptions } from './connection/db-connection-options';
