+++ packages/cli/src/auth/__tests__/auth.service.test.ts
@@ -1,4 +1,5 @@
 import type { GlobalConfig } from '@n8n/config';
+import { Time } from '@n8n/constants';
 import type { User } from '@n8n/db';
 import type { InvalidAuthTokenRepository } from '@n8n/db';
 import type { UserRepository } from '@n8n/db';
@@ -8,7 +9,7 @@ import jwt from 'jsonwebtoken';
 
 import { AuthService } from '@/auth/auth.service';
 import config from '@/config';
-import { AUTH_COOKIE_NAME, Time } from '@/constants';
+import { AUTH_COOKIE_NAME } from '@/constants';
 import type { AuthenticatedRequest } from '@/requests';
 import { JwtService } from '@/services/jwt.service';
 import type { UrlService } from '@/services/url.service';
