+++ packages/cli/src/constants.ts
@@ -1,3 +1,4 @@
+import { Time } from '@n8n/constants';
 import { readFileSync, statSync } from 'fs';
 import type { n8n } from 'n8n-core';
 import type { ITaskDataConnections } from 'n8n-workflow';
@@ -73,30 +74,6 @@ export const CREDENTIAL_BLANKING_VALUE = '__n8n_BLANK_VALUE_e5362baf-c777-4d57-a
 
 export const UM_FIX_INSTRUCTION =
 	'Please fix the database by running ./packages/cli/bin/n8n user-management:reset';
-
-/**
- * Convert time from any time unit to any other unit
- */
-export const Time = {
-	milliseconds: {
-		toMinutes: 1 / (60 * 1000),
-		toSeconds: 1 / 1000,
-	},
-	seconds: {
-		toMilliseconds: 1000,
-	},
-	minutes: {
-		toMilliseconds: 60 * 1000,
-	},
-	hours: {
-		toMilliseconds: 60 * 60 * 1000,
-		toSeconds: 60 * 60,
-	},
-	days: {
-		toSeconds: 24 * 60 * 60,
-		toMilliseconds: 24 * 60 * 60 * 1000,
-	},
-};
 
 export const MIN_PASSWORD_CHAR_LENGTH = 8;
