+++ packages/cli/src/controllers/oauth/__tests__/oauth1-credential.controller.test.ts
@@ -1,4 +1,5 @@
 import { Logger } from '@n8n/backend-common';
+import { Time } from '@n8n/constants';
 import type { CredentialsEntity } from '@n8n/db';
 import type { User } from '@n8n/db';
 import { CredentialsRepository } from '@n8n/db';
@@ -10,7 +11,6 @@ import { Cipher, type InstanceSettings, ExternalSecretsProxy } from 'n8n-core';
 import type { IWorkflowExecuteAdditionalData } from 'n8n-workflow';
 import nock from 'nock';
 
-import { Time } from '@/constants';
 import { OAuth1CredentialController } from '@/controllers/oauth/oauth1-credential.controller';
 import { CredentialsFinderService } from '@/credentials/credentials-finder.service';
 import { CredentialsHelper } from '@/credentials-helper';
