+++ packages/cli/src/modules/insights/insights-pruning.service.ts
@@ -1,8 +1,7 @@
 import { LicenseState, Logger } from '@n8n/backend-common';
+import { Time } from '@n8n/constants';
 import { Service } from '@n8n/di';
 import { strict } from 'assert';
-
-import { Time } from '@/constants';
 
 import { InsightsByPeriodRepository } from './database/repositories/insights-by-period.repository';
 import { InsightsConfig } from './insights.config';
