+++ packages/cli/src/task-runners/task-broker/__tests__/task-broker.service.test.ts
@@ -1,10 +1,10 @@
 import type { Logger } from '@n8n/backend-common';
 import type { TaskRunnersConfig } from '@n8n/config';
+import { Time } from '@n8n/constants';
 import type { RunnerMessage, TaskResultData } from '@n8n/task-runner';
 import { mock } from 'jest-mock-extended';
 import { ApplicationError, type INodeTypeBaseDescription } from 'n8n-workflow';
 
-import { Time } from '@/constants';
 import type { TaskRunnerLifecycleEvents } from '@/task-runners/task-runner-lifecycle-events';
 
 import { TaskRejectError } from '../errors/task-reject.error';
