+++ packages/cli/src/task-runners/task-runner-process-restart-loop-detector.ts
@@ -1,4 +1,5 @@
-import { Time } from '@/constants';
+import { Time } from '@n8n/constants';
+
 import { TaskRunnerRestartLoopError } from '@/task-runners/errors/task-runner-restart-loop-error';
 import type { TaskRunnerProcess } from '@/task-runners/task-runner-process';
 import { TypedEmitter } from '@/typed-emitter';
