+++ packages/cli/src/workflows/workflow-history.ee/workflow-history-manager.ee.ts
@@ -1,8 +1,7 @@
+import { Time } from '@n8n/constants';
 import { WorkflowHistoryRepository } from '@n8n/db';
 import { Service } from '@n8n/di';
 import { DateTime } from 'luxon';
-
-import { Time } from '@/constants';
 
 import {
 	getWorkflowHistoryPruneTime,
