+++ packages/cli/test/integration/executions-pruning.service.test.ts
@@ -1,13 +1,13 @@
 import { mockLogger } from '@n8n/backend-test-utils';
 import { ExecutionsConfig } from '@n8n/config';
+import { Time } from '@n8n/constants';
 import type { ExecutionEntity } from '@n8n/db';
 import { ExecutionRepository } from '@n8n/db';
+import { DbConnection } from '@n8n/db';
 import { Container } from '@n8n/di';
 import { BinaryDataService, InstanceSettings } from 'n8n-core';
 import type { ExecutionStatus, IWorkflowBase } from 'n8n-workflow';
 
-import { Time } from '@/constants';
-import { DbConnection } from '@/databases/db-connection';
 import { ExecutionsPruningService } from '@/services/pruning/executions-pruning.service';
 
 import {
