+++ packages/cli/test/integration/shared/db/workflow-statistics.ts
@@ -15,6 +15,7 @@ export async function createWorkflowStatisticsItem(
 		workflowId,
 	});
 
+	// @ts-ignore CAT-957
 	await Container.get(WorkflowStatisticsRepository).insert(entity);
 
 	return entity;
