+++ packages/cli/test/integration/shared/test-db.ts
@@ -1,12 +1,11 @@
 import { GlobalConfig } from '@n8n/config';
 import type { entities } from '@n8n/db';
+import { DbConnection } from '@n8n/db';
+import { DbConnectionOptions } from '@n8n/db';
 import { Container } from '@n8n/di';
 import type { DataSourceOptions } from '@n8n/typeorm';
 import { DataSource as Connection } from '@n8n/typeorm';
 import { randomString } from 'n8n-workflow';
-
-import { DbConnection } from '@/databases/db-connection';
-import { DbConnectionOptions } from '@/databases/db-connection-options';
 
 export const testDbPrefix = 'n8n_test_';
